Mudanças no vídeo

+++ ScreenSound.Shared.Modelos/Modelos/AvaliacaoArtista.cs
@@ -7,5 +7,5 @@ public class AvaliacaoArtista
     public int ArtistaId { get; set; }
     public virtual Artista Artista { get; set; }
     public int PessoaId { get; set; }
-    public int Nota { get; set; }
+    public double Nota { get; set; }
 }
